requirements.txt
Django>=5.2,<6.0
asgiref==3.11.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
python-dotenv==1.2.1
PyYAML==6.0.3
sqlparse==0.5.4
uritemplate==4.2.0
psycopg2-binary>=2.9
reportlab==4.0.0
